requirements.txt
async-generator==1.10
attrs==22.2.0
autopep8==2.0.2
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
google-api-core==2.11.0
google-auth==2.17.2
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
h11==0.14.0
idna==3.4
imageio==2.27.0
lazy_loader==0.2
networkx==3.1
numpy==1.24.2
outcome==1.2.0
packaging==23.0
Pillow==9.5.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PySocks==1.7.1
python-dotenv==1.0.0
PyWavelets==1.4.1
requests==2.28.2
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tifffile==2023.3.21
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
webdriver-manager==3.8.5
wsproto==1.2.0
